refactor(wallet-sdk): change file naming convention, change service names
