+++ sdk/wallet-sdk/src/wallet/namespace/amulet/types.ts
@@ -8,6 +8,24 @@ export type PreapprovalParties = {
     provider?: PartyId
 }
 
+export interface FeaturedAppService {
+    /**
+     * Looks up if a party has FeaturedAppRight.
+     * Has an in built retry and delay between attempts
+     * @returns If defined, a contract of Daml template `Splice.Amulet.FeaturedAppRight`.
+     */
+    rights: (
+        options: LookupFeaturedAppRightsOptions
+    ) => Promise<FeaturedAppRight | undefined>
+    /**
+     * Submits a command to grant feature app rights for validator operator.
+     * @returns A contract of Daml template `Splice.Amulet.FeaturedAppRight`.
+     */
+    grant: (
+        options?: GrantFeaturedAppRightsOptions
+    ) => Promise<FeaturedAppRight | undefined>
+}
+
 export type FeaturedAppRight = {
     template_id: string
     contract_id: string
