+++ sdk/wallet-sdk/src/wallet/namespace/events/namespace.ts
