--- sdk/wallet-sdk/src/wallet/namespace/ledger/dar/client.ts
@@ -1,57 +1,0 @@
-// Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { CommonCtx } from '../../../sdk.js'
-import { Ops } from '@canton-network/core-provider-ledger'
-
-export class Dar {
-    constructor(private readonly sdkContext: CommonCtx) {}
-
-    async upload(
-        darBytes: Uint8Array | Buffer,
-        packageId: string,
-        synchronizerId?: string,
-        vetAllPackages?: boolean
-    ) {
-        const isUploaded = await this.check(packageId)
-
-        if (isUploaded) {
-            this.sdkContext.logger.info(
-                { packageId },
-                'DAR already uploaded, skipping upload'
-            )
-            return
-        }
-
-        await this.sdkContext.ledgerProvider.request<Ops.PostV2Packages>({
-            method: 'ledgerApi',
-            params: {
-                resource: '/v2/packages',
-                requestMethod: 'post',
-                query: {
-                    synchronizerId:
-                        synchronizerId ?? this.sdkContext.defaultSynchronizerId,
-                    vetAllPackages: vetAllPackages ?? true,
-                },
-                body: darBytes as never,
-                headers: { 'Content-Type': 'application/octet-stream' },
-            },
-        })
-    }
-
-    async check(packageId: string): Promise<boolean> {
-        const result =
-            await this.sdkContext.ledgerProvider.request<Ops.GetV2Packages>({
-                method: 'ledgerApi',
-                params: {
-                    resource: '/v2/packages',
-                    requestMethod: 'get',
-                },
-            })
-
-        return (
-            Array.isArray(result.packageIds) &&
-            result.packageIds.includes(packageId)
-        )
-    }
-}
