+++ sdk/wallet-sdk/src/wallet/namespace/party/external/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './client.js'
+export * from './service.js'
 export * from './prepared.js'
 export * from './signed.js'
 export * from './types.js'
