+++ sdk/wallet-sdk/src/wallet/namespace/party/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './client.js'
+export * from './namespace.js'
 export * from './external/index.js'
-export * from './internal.js'
+export * from './internal/index.js'
