+++ sdk/wallet-sdk/src/wallet/namespace/party/namespace.ts
@@ -3,17 +3,17 @@
 
 import { PartyId } from '@canton-network/core-types'
 import { CommonCtx } from '../../sdk.js'
-import { ExternalParty } from './external/index.js'
-import { InternalParty } from './internal.js'
+import { ExternalPartyService } from './external/index.js'
+import { InternalPartyService } from './internal/index.js'
 import { Ops } from '@canton-network/core-provider-ledger'
 
 export default class PartyNamespace {
-    public readonly internal: InternalParty
-    public readonly external: ExternalParty
+    public readonly internal: InternalPartyService
+    public readonly external: ExternalPartyService
 
     constructor(private readonly ctx: CommonCtx) {
-        this.internal = new InternalParty(ctx)
-        this.external = new ExternalParty(ctx)
+        this.internal = new InternalPartyService(ctx)
+        this.external = new ExternalPartyService(ctx)
     }
 
     /**
