+++ sdk/wallet-sdk/src/wallet/namespace/token/transfer/index.ts
@@ -1,5 +1,5 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './client.js'
+export * from './service.js'
 export * from './types.js'
