+++ sdk/wallet-sdk/src/wallet/namespace/token/transfer/service.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { findAsset, TokenNamespaceConfig } from '../client.js'
+import { findAsset, TokenNamespaceConfig } from '../namespace.js'
 import { PartyId } from '@canton-network/core-types'
 import {
     TRANSFER_INSTRUCTION_INTERFACE_ID,
