+++ sdk/wallet-sdk/src/wallet/namespace/transactions/prepared.ts
@@ -7,7 +7,7 @@ import {
 } from '@canton-network/core-signing-lib'
 import { SignedTransaction } from './signed.js'
 import { CommonCtx } from '../../sdk.js'
-import { LedgerNamespace } from '../ledger/client.js'
+import { LedgerNamespace } from '../ledger/namespace.js'
 import { Ops } from '@canton-network/core-provider-ledger'
 import { decodePreparedTransaction } from '@canton-network/core-tx-visualizer'
